Added Piece classes & sprites
